[nc] bugfix (linter) static cast motor control enum to match operand type

+++ pdu/Src/controlStateMachine.c
@@ -406,7 +406,7 @@ uint32_t motorsOff(uint32_t event)
         ERROR_PRINT("Failed to send pdu channel status CAN message\n");
     }
 
-    if (event == MTR_EV_EM_DISABLE) {
+    if (event == static_cast<uint32_t>(MTR_EV_EM_DISABLE)) {
         return MTR_STATE_Motors_Off;
     } else {
         return MTR_STATE_Critical;
